added documentation and moved simple_module to the corect directory
