+++ src/app/models/editmodels/editable_module_input.ts
@@ -4,6 +4,9 @@ import {LearningGoal} from '../learninggoal';
 import {AssesmentPart} from '../assesment_part';
 import {PriorKnowledgeReference} from '../prior_knowledge_reference.model';
 
+// This model is used for posting an edited module to the backed.
+// This model is used in module-edit.component.ts
+// The backend service method used to post this model is updateEditableModule
 export class EditableModuleInput {
   id: number;
   code: string;
@@ -35,10 +38,12 @@ export class EditableModuleInput {
     this.credentials = output.credentials;
     this.project_flag = output.project_flag;
     this.prior_knowledge_references = output.prior_knowledge_references;
-    this.lecturers = new Array();
     this.learning_goals = output.learning_goals;
     this.assesment_parts = output.assesment_parts;
     this.prior_knowledge_references = output.prior_knowledge_references;
+
+    // Giving meaning to a empty object
+    this.lecturers = new Array();
     for (let i = 0; i < output.active_lecturers.length; i++) {
       this.lecturers[i] = output.active_lecturers[i].id;
     }
