+++ src/app/models/editmodels/simple_module.ts
@@ -1,0 +1,6 @@
+// This simplified module model exists to make posting information easier used in the module-edit.component.ts
+export interface SimpleModule {
+  id: number;
+  code: string;
+  name: string;
+}
