+++ src/app/models/qualificiationfiltermodels/qualifications_learning_goal.model.ts
@@ -1,3 +1,4 @@
+// A minimum version of a learning goal.
 export interface QualificationsLearningGoal {
   name: string;
   description: string;
