+++ src/app/models/qualificiationfiltermodels/qualifications_overview_semester.model.ts
@@ -1,5 +1,6 @@
 import {QualificationsModule} from './qualifications_module';
 
+// This model exists because there was need of a modified module model.
 export interface  QualificationsOverviewSemester {
   semester: number;
   qualifications_modules: QualificationsModule[];
