--- src/app/models/qualificiationfiltermodels/simple_module.ts
@@ -1,5 +1,0 @@
-export interface SimpleModule {
-  id: number;
-  code: string;
-  name: string;
-}
